src/python_services/asr/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
openai-whisper>=20230314
torch>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0
pydantic>=2.0.0
